apps/backend/pyproject.toml
[project]
name = "gigaboard-backend"
version = "0.1.0"
description = "GigaBoard AI Analytics Platform - Backend"
requires-python = ">=3.11"
authors = [
    { name = "GigaBoard Team", email = "dev@gigaboard.local" }
]

dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.0",
    "python-socketio==5.11.2",
    "redis==5.0.1",
    "pydantic[email]>=2.10",
    "pydantic-settings==2.1.0",
    "SQLAlchemy==2.0.36",
    "asyncpg==0.30.0",
    "python-dotenv==1.0.1",
    "langchain==0.3.7",
    "langchain-gigachat==0.3.12",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.2.1",
    "python-jose[cryptography]==3.3.0",
    "PyJWT==2.10.1",
    "cryptography==41.0.7",
    "alembic==1.13.1",
    "asyncpg==0.30.0",  # PostgreSQL async driver for main app
    "pandas==2.2.0",
    "httpx==0.26.0",
    "python-multipart>=0.0.6",
    "ddgs>=1.0.0",
    "chardet==5.2.0",  # Encoding detection
    # Extractors dependencies
    "openpyxl==3.1.2",  # Excel support
    "aiomysql==0.2.0",  # MySQL async support
    "pyarrow==14.0.1",  # Parquet support
    # Database connection drivers (for data extraction only, not for main app DB)
    "pymysql==1.1.0",  # MySQL
    "clickhouse-driver==0.2.9",  # ClickHouse
    "pymongo==4.10.1",  # MongoDB
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
    "httpx==0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
# Указываем корневой .venv через переменную окружения UV_PROJECT_ENVIRONMENT
# или используем: uv run --python ../../.venv/Scripts/python.exe
